affichage des boutons en fonction du role

+++ includes/_createArticle.php
@@ -1,10 +1,7 @@
-
-
-
 <article>
     <header>
         <img src="<?php echo $product['image_url'] ?>" alt="" />
-        <p class="type">type:                              <?php echo $product['transaction_type'] ?></p>
+        <p class="type">type: <?php echo $product['transaction_type'] ?></p>
     </header>
     <div class="title">
         <h3><?php echo $product['title'] ?></h3>
@@ -14,13 +11,15 @@
     <p class="description">
         <?php echo $product['description'] ?>
     </p>
-
+    <?php if (isset($_SESSION['id'])): ?>
     <div id=button-container>
         <a href="updateFavoris.php?id=<?php echo $product['id'] ?>" class="link">Favori</a>
-       <!-- <?php echo $isFavorited ? '❤️' : '🤍'; ?> -->
-        <a href="delete.php?id=<?php echo $product['id'] ?>" class="link">Supprimer</a>
-        <a href="update.php?id=<?php echo $product['id'] ?>" class="link">Modifier</a>
+        <?php if (($user_role === "admin" || $user_id === $product['user_id'])): ?>
+            <a href="delete.php?id=<?php echo $product['id'] ?>" class="link">Supprimer</a>
+            <a href="update.php?id=<?php echo $product['id'] ?>" class="link">Modifier</a>
+        <?php endif; ?>
     </div>
+    <?php endif; ?>
     <button>Contact</button>
 
 </article>
